Lazy Entity Fetch, OptionPaneMessages, RoseException, Logging++

+++ src/bn/blaszczyk/roseapp/controller/GUIController.java
@@ -5,11 +5,15 @@ import java.util.ArrayList;
 import java.util.List;
 
 import javax.swing.JDialog;
+import javax.swing.JOptionPane;
+
+import org.apache.log4j.Logger;
 
 import bn.blaszczyk.rose.model.Identifyable;
 import bn.blaszczyk.rose.model.Readable;
 import bn.blaszczyk.rose.model.RelationType;
 import bn.blaszczyk.rose.model.Writable;
+import bn.blaszczyk.roseapp.RoseException;
 import bn.blaszczyk.roseapp.tools.EntityUtils;
 import bn.blaszczyk.roseapp.tools.Preferences;
 import bn.blaszczyk.roseapp.tools.TypeManager;
@@ -22,6 +26,8 @@ import bn.blaszczyk.roseapp.view.panels.crud.StartPanel;
 import bn.blaszczyk.roseapp.view.panels.settings.SettingsPanel;
 
 public class GUIController {
+	
+	private static final Logger LOGGER = Logger.getLogger(GUIController.class);
 
 	private ModelController modelController;
 	private MainFrame mainFrame;
@@ -132,6 +138,7 @@ public class GUIController {
 	
 	public void addTab( RosePanel panel, String name, String iconFile)
 	{
+		LOGGER.info("open tab: \"" + name + "\" : " + panel );
 		for(ActionPack a : actionPacks)
 			panel.addRoseListener(a);
 		int index = getObjectsTabIndex(panel.getShownObject());
@@ -169,7 +176,8 @@ public class GUIController {
 	
 	public void viewCurrent()
 	{
-		openEntityTab( ((Readable) mainFrame.getSelectedPanel().getShownObject()), false );
+		if(confirmDialog("Discard changes?", "There are unsaved changes"))
+			openEntityTab( ((Readable) mainFrame.getSelectedPanel().getShownObject()), false );
 	}
 	
 	public void saveCurrent()
@@ -179,13 +187,30 @@ public class GUIController {
 		if(panel instanceof FullEditPanel)
 		{
 //			openEntityTab((Readable) panel.getShownObject(),false);
-			modelController.commit();
+			try
+			{
+				modelController.commit();
+			}
+			catch (RoseException e)
+			{
+				e.printStackTrace();
+				LOGGER.error("Error saving " + panel.getShownObject(), e);
+				errorDialog(e, "Save Error");
+			}
 		}
 		notifyListeners();
 	}
 	
 	public void closeCurrent()
 	{
+		if(mainFrame.getSelectedPanel().hasChanged())
+		{
+			int option = yesnocancelDialog("Save before closing?", "There are unsaved changes.");
+			if(option == JOptionPane.CANCEL_OPTION)
+				return;
+			if(option == JOptionPane.YES_OPTION)
+				saveCurrent();
+		}
 		mainFrame.closeCurrent();
 		notifyListeners();
 	}
@@ -201,14 +226,31 @@ public class GUIController {
 	{
 		RosePanel c = mainFrame.getSelectedPanel();
 		if( c instanceof RosePanel && ((RosePanel)c).getShownObject() instanceof Writable )
-			openEntityTab( modelController.createCopy( (Writable) ((RosePanel)c).getShownObject() ), true );
+			try
+			{
+				openEntityTab( modelController.createCopy( (Writable) ((RosePanel)c).getShownObject() ), true );
+			}
+			catch (RoseException e)
+			{
+				e.printStackTrace();
+				errorDialog(e, "Error");
+			}
 		notifyListeners();
 	}
 
 	public <T extends Writable> void openNew(Class<T> type)
 	{
-		openEntityTab( modelController.createNew( type ), true );
-		notifyListeners();
+		try
+		{
+			openEntityTab( modelController.createNew( type ), true );
+			notifyListeners();
+		}
+		catch (RoseException e)
+		{
+			e.printStackTrace();
+			LOGGER.error("Error creating new " + type.getName(), e);
+			errorDialog(e, "Error");
+		}
 	}
 
 	public void openNew()
@@ -229,11 +271,22 @@ public class GUIController {
 
 	public void addNew(Writable entity, int index)
 	{
-		Writable subEntity = modelController.createNew( entity.getEntityClass(index).asSubclass(Writable.class) );
-		entity.addEntity(index, subEntity);
-		modelController.update(entity,subEntity);
-		openEntityTab( subEntity, true);
-		notifyListeners();
+		if(entity == null)
+			return;
+		try
+		{
+			Writable subEntity = modelController.createNew( entity.getEntityClass(index).asSubclass(Writable.class) );
+			entity.addEntity(index, subEntity);
+			modelController.update(entity,subEntity);
+			openEntityTab( subEntity, true);
+			notifyListeners();
+		}
+		catch (RoseException e)
+		{
+			e.printStackTrace();
+			LOGGER.error("Error adding new entity field at index " + index + " to\r\n" + EntityUtils.toStringFull(entity) , e);
+			errorDialog(e, "Error");
+		}
 	}
 
 	public void saveAll()
@@ -246,23 +299,45 @@ public class GUIController {
 //			if( mainFrame.getPanel(i) instanceof FullEditPanel )
 //				openEntityTab((Readable) panel.getShownObject(),false);
 			}
-		modelController.commit();
+		try
+		{
+			modelController.commit();
+		}
+		catch (RoseException e)
+		{
+			e.printStackTrace();
+			LOGGER.error("Save error", e);
+			errorDialog(e, "Save Error");
+		}
 		mainFrame.setSelectedIndex(current);
 		notifyListeners();
 	}
 
 	public void delete(Writable entity)
 	{
-		for(RosePanel panel : mainFrame)
+		if(entity == null)
+			return;
+		if(! confirmDialog("Delete " + entity + "?", "Confirm Delete"))
+			return;
+		try
 		{
-			if(panel.getShownObject() instanceof Readable )
-				if( ((Readable)panel.getShownObject()).equals(entity))
-					mainFrame.removePanel(panel);
-			if(panel.getShownObject().equals(entity.getClass()))
-				panel.refresh();
+			modelController.delete(entity);
+			for(RosePanel panel : mainFrame)
+			{
+				if(panel.getShownObject() instanceof Readable )
+					if( ((Readable)panel.getShownObject()).equals(entity))
+						mainFrame.removePanel(panel);
+				if(panel.getShownObject().equals(entity.getClass()))
+					panel.refresh();
+			}
+			notifyListeners();
 		}
-		modelController.delete(entity);
-		notifyListeners();
+		catch (RoseException e)
+		{
+			e.printStackTrace();
+			LOGGER.error("Error deleting " + EntityUtils.toStringFull(entity), e);
+			errorDialog(e, "Error deleting " + entity);
+		}
 	}
 
 	public void closeAll()
@@ -275,7 +350,7 @@ public class GUIController {
 	public void deleteOrphans()
 	{
 		for(Class<? extends Readable> type : TypeManager.getEntityClasses())
-			for(Readable entity : modelController.getAllEntites(type))
+			for(Readable entity : modelController.getEntites(type))
 			{
 				boolean orphan = true;
 				for(int i = 0; i < entity.getEntityCount(); i++)
@@ -305,5 +380,25 @@ public class GUIController {
 		if(out.getEntityCount() > in.getEntityCount())
 			return out;
 		return in;
+	}
+	
+	/*
+	 * messages
+	 */
+	
+	private int yesnocancelDialog(String message, String title)
+	{
+		return JOptionPane.showConfirmDialog(mainFrame, message, title, JOptionPane.YES_NO_CANCEL_OPTION, JOptionPane.QUESTION_MESSAGE);
+	}
+	
+	private void errorDialog(RoseException e, String title)
+	{
+		JOptionPane.showMessageDialog(mainFrame, e.getFullMessage(), title, JOptionPane.ERROR_MESSAGE);
+	}
+	
+	private boolean confirmDialog(String message, String title)
+	{
+		return JOptionPane.showConfirmDialog(mainFrame, message, title, JOptionPane.OK_CANCEL_OPTION, JOptionPane.QUESTION_MESSAGE)
+				== JOptionPane.OK_OPTION;
 	}
 }
